requirements.txt
groq
python-dotenv
streamlit
aiohttp